Added insync method overloads to all values that are arrays
If they're set to nil or empty(they're always empty, not nil, due to the way the
array_matching all is done) they will be checked against the IS value to see if it
is nil, if it is, do nothing, else, suggest to provider that the value has changed.

That way we can comment out values or set them to [] and it will notice it should flush
and write zonefile.

Default insync method checks and if .empty? it returns true and does not submit for
flush and write, not what we want.